backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.7.0
reportlab>=4.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
